Make space between text for cursor + cut and copy for whole lines

+++ src/gui/item_textfield.hpp
@@ -63,7 +63,7 @@ protected:
 
   const char m_cursor_char;
   const std::string m_cursor_char_str;
-  const int m_cursor_char_width;
+  const float m_cursor_char_width;
   int m_cursor_left_offset;
 
 private:
